[Obs Onboarding] Migrate API test to deployment agnostic framework
