+++ x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.index.ts
@@ -23,5 +23,6 @@ export default function ({ loadTestFile }: DeploymentAgnosticFtrProviderContext)
     loadTestFile(require.resolve('../../apis/observability/synthetics'));
     loadTestFile(require.resolve('../../apis/observability/ai_assistant'));
     loadTestFile(require.resolve('../../apis/observability/streams'));
+    loadTestFile(require.resolve('../../apis/observability/onboarding'));
   });
 }
