+++ x-pack/test/api_integration/deployment_agnostic/configs/stateful/oblt.index.ts
@@ -17,5 +17,6 @@ export default function ({ loadTestFile }: DeploymentAgnosticFtrProviderContext)
     loadTestFile(require.resolve('../../apis/observability/infra'));
     loadTestFile(require.resolve('../../apis/observability/ai_assistant'));
     loadTestFile(require.resolve('../../apis/observability/streams'));
+    loadTestFile(require.resolve('../../apis/observability/onboarding'));
   });
 }
