+++ x-pack/test/observability_onboarding_api_integration/tests/logs/privileges.spec.ts
@@ -21,30 +21,10 @@ export default function ApiTest({ getService }: FtrProviderContext) {
 
   registry.when('Api Key privileges check', { config: 'basic' }, () => {
     describe('when missing required privileges', () => {
-      it('returns false when user has reader privileges', async () => {
-        const privileges = await callApiAs('readUser');
-
-        expect(privileges.body.hasPrivileges).not.ok();
-      });
-
       it('returns false when user has no access privileges', async () => {
         const privileges = await callApiAs('noAccessUser');
 
         expect(privileges.body.hasPrivileges).not.ok();
-      });
-    });
-
-    describe('when required privileges are set', () => {
-      it('returns true when user has logMonitoring privileges', async () => {
-        const privileges = await callApiAs('logMonitoringUser');
-
-        expect(privileges.body.hasPrivileges).ok();
-      });
-
-      it('returns true when user has admin privileges', async () => {
-        const privileges = await callApiAs('adminUser');
-
-        expect(privileges.body.hasPrivileges).ok();
       });
     });
   });
